+++ src/vs/workbench/services/extensions/worker/extensionHostWorker.ts
@@ -10,6 +10,7 @@ import { isMessageOfType, MessageType, createMessageOfType } from 'vs/workbench/
 import { IInitData } from 'vs/workbench/api/common/extHost.protocol';
 import { ExtensionHostMain } from 'vs/workbench/services/extensions/common/extensionHostMain';
 import { IHostUtils } from 'vs/workbench/api/common/extHostExtensionService';
+import { NestedWorker } from 'vs/workbench/services/extensions/worker/polyfillNestedWorker';
 import * as path from 'vs/base/common/path';
 
 import 'vs/workbench/api/common/extHost.common.services';
@@ -58,101 +59,9 @@ if ((<any>self).Worker) {
 	};
 
 } else {
-
-	function _bootstrapFn(workerUrl: string) {
-
-		const listener: EventListener = (event: Event): void => {
-			// uninstall handler
-			self.removeEventListener('message', listener);
-
-			// get data
-			const port = (<MessageEvent>event).data;
-
-			// postMessage
-			Object.defineProperty(self, 'postMessage', {
-				value(data: any, transferOrOptions?: any) {
-					port.postMessage(data, transferOrOptions);
-				}
-			});
-
-			// onmessage
-			let onmessage: Function | null = null;
-			Object.defineProperty(self, 'onmessage', {
-				get() { return onmessage; },
-				set(value) { onmessage = value; },
-			});
-			port.onmessage = function (msg: MessageEvent) {
-				self.dispatchEvent(new MessageEvent('message', { data: msg.data }));
-				if (typeof onmessage === 'function') {
-					onmessage(msg.data);
-				}
-			};
-
-			// load module
-			importScripts(workerUrl);
-		};
-
-		self.addEventListener('message', listener);
-	}
-
-	// NO support for workers from worker
-	(<any>self).Worker = class extends EventTarget implements Worker {
-
-		onmessage: ((this: Worker, ev: MessageEvent<any>) => any) | null = null;
-		onmessageerror: ((this: Worker, ev: MessageEvent<any>) => any) | null = null;
-		onerror: ((this: AbstractWorker, ev: ErrorEvent) => any) | null = null;
-
-		readonly terminate: () => void;
-		readonly postMessage: (message: any, options?: any) => void;
-
+	(<any>self).Worker = class extends NestedWorker {
 		constructor(stringOrUrl: string | URL, options?: WorkerOptions) {
-			super();
-
-			// create bootstrap script
-			const bootstrap = `((${_bootstrapFn.toString()})('${stringOrUrl}'))`;
-			const blob = new Blob([bootstrap], { type: 'application/javascript' });
-			const blobUrl = URL.createObjectURL(blob);
-
-			const channel = new MessageChannel();
-			const id = blobUrl; // works because blob url is unique, needs ID pool otherwise
-
-			nativePostMessage({
-				type: '_newWorker',
-				id,
-				port: channel.port2,
-				url: blobUrl,
-				options: { name: path.basename(String(stringOrUrl)), ...options },
-			}, [channel.port2]);
-
-			// worker-impl: functions
-			this.postMessage = channel.port1.postMessage.bind(channel.port1);
-			this.terminate = () => {
-				channel.port1.postMessage({ type: '_terminateWorker', id });
-				URL.revokeObjectURL(blobUrl);
-			};
-
-			// worker-impl: events
-			let _onmessage: ((ev: MessageEvent<any>) => any) | null = null;
-			Object.defineProperties(this, {
-				'onmessage': {
-					get() {
-						return _onmessage;
-					},
-					set(value: ((ev: MessageEvent<any>) => any) | null) {
-						_onmessage = value;
-					}
-				}
-				// TODO - define error, messageerror event
-			});
-
-			channel.port1.onmessageerror = evt => this.dispatchEvent(evt);
-			channel.port1.onmessage = (evt) => {
-				const msgEvent = new MessageEvent('message', { data: evt.data });
-				if (_onmessage) {
-					_onmessage(msgEvent.data);
-				}
-				this.dispatchEvent(msgEvent);
-			};
+			super(nativePostMessage, stringOrUrl, { name: path.basename(stringOrUrl.toString()), ...options });
 		}
 	};
 }
